db.get batch at isolate layer (#22802)
allow async syscalls that are queued up together to potentially be run in a batch, with a new `AsyncSyscallBatch` enum.

then thread the batch down through isolate and Transaction, but not TransactionIndex because i'd like for #22711 to land first and simplify the fetch at that layer.

note that instead of returning a single error for the whole batch, we return an error for each item in the batch. This way we can independently resolve or reject each promise.

GitOrigin-RevId: cdda682f224719e3026fffe17554e5ecbcf920e6