Clean up fetch client. (#23167)
With fetch client in runtime... we can't really configure the site per request, which is needed in Funrun. In general, relying on env var so deep into the stack is not a good idea.

Pass a fetch client to actions instead of using the one from runtime. This allow us to create fetch client per request in Funrun. I am going to remove fetch() from runtime altogether in a separate PR.

GitOrigin-RevId: 717089548a29af218b4ad6a6245c3657d7ef04d8