Pull out `Transaction::*_user_facing` methods into `UserFacingModel` (#23270)
This is a smaller version of #22531 that pulls out the `user_facing` methods into a `UserFacingModel` struct within `database/src/model`. This model isn't required for bootstrap, so I didn't put it in `bootstrap_model`. We already have the pattern of models calling into other models, so I decided to reuse that rather than introduce the new concept of a transaction "view."

I'm going to land this since it touches many files, but please leave any feedback you have!

GitOrigin-RevId: 31e3572400d7c3de47bc5117219582bd92a1369f