Add gauages to track when index/database workers are doing useful work (#23247)
While we have timers for most of these workers, they don't show you
what's currently running. So if a worker runs for a long time you can't
be sure of its status until it finishes, which isn't super helpful.

The gauges will also show us when workers are or aren't running
concurrently, which might also make it easier to understand how they
interact with each other or other user actions.

I wanted this specifically for search index workers, but I figure the
concept is useful for most other types of background workers, so I've
added a similar to gauge to the ones I could find.

GitOrigin-RevId: f7cd20d2da115923ec9454b465199baba544139c