Bootstrap vector indexes in one pass over document log (#23272)
This PR changes the algorithm for bootstrapping vector indexes.
Before: for every vector index we would walk the document log since it was written to disk to load updates.
Now: We find the min timestamp of all the vector indexes that we'd need to walk the document log from, and walk the document log once, checking which vector indexes we need to update on each revision.

The first two commits are just cleaning up some code that could be simpler.

GitOrigin-RevId: 65307bac3b7681187ea52bc11911c85889c45df7