Don't drop log_line_receiver (#23307)
When we log in node functions, we send something down the `log_line_sender`, even though we ignore these for analyze (or anything other than running a function). However, by dropping the receiver with `_`, it means that the send fails, which fails the entire request. Instead, `_log_line_receiver` delays dropping this until it goes out of scope (which is after we've received the full response)

GitOrigin-RevId: 00b322967e80185a39a84010d89397f0ac70d99a