Expect overloaded errors in fast forward worker (#23277)
Vector indexes (and soon search indexes) will throw overloaded errors if we interact with them while they're bootstrapping. The index workers do this reliably because they run as soon as the backend is started. So we should be expecting and not reporting overloaded errors below some threshold.

This fixes a logging regression with the refactor to create fast forward worker. Previously we'd have ignored these for vector indexes because the vector index worker used this logic for both fast forwarding and building.

The only behavior change here is that we will report fewer overloaded errors to sentry.

GitOrigin-RevId: cc629259dd4f995d23f0118944a4c7ff7149bc8b