[DatabaseIndexSnapshot] factor out a method that can run in parallel (#23415)
this is a refactor, not changing behavior.

instead of populating the cache as we iterate over results, we populate the cache at the end, after we have all results.
this allows us to split out a method `fetch_cache_misses` which interacts with the cache only through arguments and return values.
thus `fetch_cache_misses` can be run in parallel with itself.

GitOrigin-RevId: 9d03269cf71092f5207b5fbeaeee6bfc5a882c23