ensure Interval advances (#23440)
return errors if we are fetching a page from an index range and the remaining Interval to query does not shrink. Do this at the lowest levels where we compute the remaining_intervals -- once where we merge it with pending writes, and once where we truncate results based on length and bytes size. and also do this at the highest level where we absolutely require the interval to be advancing -- in the Query loop.

GitOrigin-RevId: 8e74577b5a53314c066a3ebaae5f2e9e54a49bdc