fix Interval not advancing (#23462)
pending deletes can make an `index_range` return no results, which is fine. but even in these cases we want to shrink the interval being read for the next page. otherwise we end up in an infinite loop.

I considered just fixing `interval_unread`, but we actually need the `interval_read` as well, to record the read set (in this case where there's an empty page). And we can't compute `interval_read` with the information available. So instead of passing `interval_unread` (i.e. `remaining_interval`) as a return value, we should return a `CursorPosition` from which we can easily derive both `interval_unread` and `interval_read`.

added regression test which fails on main.

GitOrigin-RevId: 5a37f0b9ef035dd59df88742afd3c56db0b9b420