Bootstrap search with vector indexes (#23377)
This change is part 1 to bootstrapping search with vector indexes. We find the earliest timestamp we need to walk the document log from, and then we bootstrap search and vector indexes in the same pass. This isn't hooked up into actually doing this async yet, just wanted to break out this part of the change so it's easier to review. I'll merge once both parts are good to go.

GitOrigin-RevId: 612ad2c225f52a9126fd9f26740d93c86e9f2464