Make search bootstrap async (#23436)
This PR makes bootstrapping search indexes async. In #23377, we added code to bootstrap search indexes with vector indexes. In this change, we link it together, adding bootstrapping state to the `SearchIndexManager` and sending the results of bootstrapping to the committer. I also moved the tests to `search_and_vector_bootstrap` and removed the standalone search bootstrap code 🎉

GitOrigin-RevId: 97f3a5d2c5e72733602d5b5b8a7abf17bf7b5d8c