Move max_isolate_workers outside of IsolateConfig (#23499)
max_isolate_workers being a property of IsolateConfig is very confusing. IsolateConfig is meant to configure a single worker.

There are few places in code where we set it, but the value from IsolateConfig is not actually used, which makes things very confusing and error prone. For example, FunctionRunnerCore sets UDF_ISOLATE_MAX_EXEC_THREADS but this is not actually used.

GitOrigin-RevId: 80b400795de2a580619520448441cfe336210de8