Revert "Revert "Make search bootstrap async" (#23530)" (#23600)
This PR undoes a revert of two PRs that make bootstrapping search indexes async. We reverted those PRs because roll out failed because we called `update` on in memory search and vector indexes for revisions prior to the `min_ts`. This PR returns early with a no-op when we call `update` on revisions prior to `min_ts` and adds tests.

GitOrigin-RevId: dd88f9e70389ed22e440e1d635bba3b0c00cf6c8