Make search bootstrap async (#23647)
Take 3 at landing asynchronous search index bootstrapping. This time I fixed a bug where we were writing successor timestamps on every revision when we added updates that happened after bootstrapping but before committing. I added a test that uses the pause client to write after we finish bootstrapping but before the worker sends the message to the committer.

Reverts #23615

GitOrigin-RevId: 557eee59d491a71a74139133bbd812296310bc6d