Add new stream_function_logs endpoint (#23612)
This endpoint will eventually be used to allow showing live log lines for actions in the dashboard and via `npx convex logs`

It returns "Completion" and "Progress" events, where "Progress" events contain log lines. Log lines from a function might appear in the "Completion" (in the case of queries and mutations) or in the "Progress" events (in the case of actions), but will not appear twice.

GitOrigin-RevId: aa5fa0965fad1f2dbb98f6c9911533f82688416b