reduce memory usage in TableIterator (#23771)
improve memory usage when using TableIterator on a large table with large documents at a timestamp that may be hours old.

1. currently TableIterator stores full documents in `skipped_keys` which can persist for the lifetime of the iterator. This causes the following memory pattern, as documents modified recently are stored until their keys are accessed by the iterator.
![Screenshot 2024-03-21 at 12 34 27 PM](https://github.com/get-convex/convex/assets/4319355/04fbd8f9-050b-4759-8ada-4ce3f2171f05)
to fix this, I changed `skipped_keys` to only store IDs, not full documents. When we need the full documents to be yielded as part of a page of results, we re-fetch. This could potentially slow down the table iterator, but I think it's worth it.

2. currently TableIterator fetches `previous_revisions` when calculating skipped_keys and stores them all in memory at once before filtering based on index key. If you start a new TableIterator several hours after the snapshot timestamp, this means we're reading hours of documents into memory before discarding them all. And you get this memory pattern:
![Screenshot 2024-03-21 at 12 38 33 PM](https://github.com/get-convex/convex/assets/4319355/e2452c19-a7af-427e-b448-22de931ff7f6)
which eventually OOMs db-verifier.
to fix this, i changed `load_revisions_at_snapshot_ts` to return a stream and chunk internally.

GitOrigin-RevId: dc3c5c9173a6bb1cb12b1c0bbb4392a5b571a368