Reenable scheduled jobs tests with `PauseClient` (#23745)
This PR fixes some flakey scheduled jobs tests by adding a `PauseClient` to the `ScheduledJobExecutor`. We wait for the job to complete before testing assertions.
Unfortunately this doesn't work for `test_cancel_recursively_scheduled_job` because we're not relying on the scheduler at all. Still investigating that one, but wanted to make some incremental progress here.

GitOrigin-RevId: c3dc398305c2ba4edba1bf729514120ace763822