Move scheduling out of the inner scheduled jobs loop (#23750)
This change more cleanly separates the responsibilities of the two
loops. The outer loop handles polling and subscriptions and logging.
The inner loop handles the mechanics of actually running jobs. The
primary goal of the change is to avoid having to write polling logic in
the inner loop.

As a side bonus this also lets us skip running the jobs query at all if
we're running jobs at our concurrency limit and are woken up due to a
subscription change. Not a big deal, but it's nice to make the super
active backend case less busy.

GitOrigin-RevId: 58b9f529701f55e94cf5c938aa17c5b698a017e8