Add a rate limiter for node actions (#23699)
Add a per backend knob for rate limiting node actions.

I've also generalized a few of the application function runner metrics to include the module environment. I'm assuming from the APPLICATION_FUNCTION_RUNNER prefixes on the metric names that we're ok with using these metrics for functions that run outside of FunctionRunner implementations.

GitOrigin-RevId: 51c5d3327ef2dcc353a1193f46d113c1e1e147e2