Only get previous revisions during search and vector bootstrapping for documents in tables with indexes (#23831)
This PR makes search and vector bootstrapping more efficient by filtering out documents that are not in tables that have indexes before getting previous revisions. I refactored to pass a `BootstrappedSearchAndVectorIndexes` struct that includes the set of tables with indexes so we can apply the same logic in the committer update. Also added metrics for the committer update, since technically search and vector search are down until the committer update is complete.

GitOrigin-RevId: ca70695094ebaa9c26f86e8f61aa94cfa2c09644