Rename RequestContext to ExecutionContext (#23870)
Rename RequestContext to ExecutionContext. As it became crystal clear recently, the struct makes sense in the context of a function execution and not as request in general. It should be generated as close to execution as possible.

I have renamed the struct, file and proto. I kept RequestId in the same file but reexported it as common::RequestId, so it is easier to move to a different file later.

GitOrigin-RevId: 67856b99bedaa28b82e8567ad3f2788c749cf915