Refactor search and vector bootstrap to be more readable (#23842)
This PR moves the `IndexesToBootstrap` creation fn and bootstrap fn to be on `IndexesToBootstrap` instead of in `SearchAndVectorBootstrapWorker`. I think it's a minor improvement to separate the code so the worker is really just the async loop and sending the message to the committer. `IndexesToBootstrap` is the where most of the logic of transforming index documents into index managers lives. This change could make it easier to unit test in the future. Right now all the tests are at the worker level.

GitOrigin-RevId: 8c61b038f324e8ca28cf402fd750609f2b6bb002