Start logging structured log lines (#23391)
We've landed the code to handle reading either structure or unstructured log lines (always persisting them as unstructured).

So now it should be safe to start logging structured log lines!

For now, the only user visible change would be the timestamps used in log streams, but pretty soon we can surface these on the dashboard or in the CLI (e.g. show timestamps per log line on the dashboard logs page)

GitOrigin-RevId: 426d238f77f0bd9cfa657005f36551778b6f0f94