Fix for enforcing document retention (#23587)
This PR enforces document retention on `load_documents` by adding a method to the `RetentionValidator` that checks if reading at a `ts` is valid.

To make this work with the deleter for index retention, we ensure that the invariant `index_retention_confirmed_deleted_ts` >= `min_documents_snapshot_ts` is always true.

Previously, this caused a bug because retention was loading documents outside of the valid `documents` snapshot window. Enforcing the above invariant will prevent this + keep us from deleting documents before the indexes are deleted.

This will be followed up by enforcing retention on `previous_revisions`

GitOrigin-RevId: cffead60258c67df3a2ca31213536bca285f0ec0