Add retention running flag to BackfillIndexState (#24069)
Create a new index_created_lower_bound field so we can update index entry in backfilled state. Currently, we rely on it never updating. Also add retention_running flag so we can so the period we don't run retention for in newly backfilled indexes is minimal.

This change adds and populates the new fields and adds a migration to backfill them properly. I am going to change IndexWorker in a separate PR.

GitOrigin-RevId: 1b28ed0391f203a373ad4706cfcc65b2628eece2