Output full bundle with `debugBundlePath` instead of a single file (#24169)
The end goal is to be able to have a documented guide for debugging large bundle size.

Without these changes, this flag only works on `deploy` (not on `dev`) and emits a single file of JSON metadata, including all the bundled files.

With these changes, we fill the provided directory with equivalent JSON metadata in `fullConfig.json`, but then also make a `.js` and `.js.map` file for all the bundled files. It also adds the flag (still hidden) to `dev`

GitOrigin-RevId: 47fbc3f57e32ee4f63be3c2d6923ea6f3dd8cc84