support `_id` filters without typescript (#24298)
support index filters on `_id` fields. don't change typescript yet so we can do a soft launch, testing in prod a bit before exposing it to customers.

the tricky part here is when it's a virtual table. The `_id` strings use the virtual table number but the underlying Intervals use the physical table number. Luckily, base32 preserves order, so ordering is all the same. When the `_id` filter is a valid virtual ID, it's easy to map it to the corresponding physical ID. But what if the `_id` filter isn't a virtual ID? Then we need to find a string in the physical ID space that corresponds to the same place in the order. Added a bunch of proptests so I think it's working.

GitOrigin-RevId: 6677664011760c805d32d87851c470c9d84876f9