make `1.0/get` syscall use CompiledQuery (#23969)
change `Transaction::get_batch` usage in syscalls to use `DeveloperQuery`. This makes it more true when we say "db.get is a special case of db.query".
By representing them as DeveloperQuery, it unblocks running `db.query` and `db.get` in parallel with each other, which will be a follow up to this PR.

Note there is a side effect: it allows developers to filter on `_id` in index filters, although without type support.

GitOrigin-RevId: 6ffaf824335dff6f15f5f8a03aaede1224a0e2e8