Get the first test `basic::test_basic` working in isolate v2 (#24563)
This PR moves the isolate v2 demo into a test and adds support for arguments, return values, and the first few ops.

Now that stuff generally works but is missing a bunch of functionality, I'm going to next push isolate v2 into `UdfTest` and have it speculatively run tests on both implementations.

GitOrigin-RevId: 5029e58aa74e66f1ee29ead6e323bff7ca282b07