Instrument acquring sempahore. (#24633)
Track how much CPU time each client_id is consuming. Since the concurrency semaphore can theoretically be held for 10 minute, it is not enough to only log when the semaphore is dropped. We also log every 10 seconds for ongoing requests.

All this will be much simpler after isolate_v2 refactor. For now, I need to track and log via a central spawned future. That code will go away but we need in the meantime.

GitOrigin-RevId: 6f6518a7738d73e229a8953717ea7486fdd816fd