[CX-6208] batch virtual table queryStreamNext (#24732)
batch `db.system.query` requests that are awaited in parallel. effectively this pulls the fetch out of `VirtualTable::index_range` up a layer so it can run in parallel with other `db.get`s and `db.query`s. Then we keep track of whether it's a virtual table in `virtual_table_versions` and map virtual documents after fetching them.

GitOrigin-RevId: 81ced12a9a69b8eeeb6fa66d366d200e8a8873de