parallel storage.getUrl (#24821)
allow `ctx.storage.getUrl` in queries and mutations to run in parallel with itself, as a batch.

note this removes a couple of the sanity checks, because i don't think they're necessary and with batching they're harder: (1) removes the assertion that there's at most one stored file with a given storage id, (2) removes the assertion that `entry.storage_id == storage_id` when `entry` was fetched with an equality index filter on `storage_id`

i wanted to make it also run in parallel with `db.get` and `db.query` but that turns out to be tricky because those need to use `DeveloperQuery` while file storage urls needs to use `ResolvedQuery`.

GitOrigin-RevId: 41e686efe0b36c60d9e91a99a5cfc8bc4d3f56d3