Reduce clones building search short lists (#24834)
This is a bit awkward because we want only the new shortlist ids out of the builder, but we want to match those shortlist ids to the term ids which the builder doesn't care about. My previous version of this method handled that by returning a result for each individual term match, but that a map look up and a clone per match.

So I've taken another stab at this by changing `add_match` to `add_matches`. `add_matches` avoids the clones and repeated lookups by returning a `vec[]` with a `Some` or `None` shortlist id for each query term, where only the new shortlist ids are some. It's a little more confusing to read, but not dramatically worse...

GitOrigin-RevId: bacd7a8651d3ab2a8a8499a50172e3471c75fba0