cancel snapshot imports that take >1 day (#25005)
we need to time bound snapshot imports. ideally they would finish quickly, but if they're taking more than a day then that's much too long. we need to be able to delete the file from S3 and the hidden tables being written to.

reporting this as a bad request is somewhat incorrect, but it does represent what we should do with the error -- send it to the customer if they're watching and cancel the import. It makes the "import will execute to completion in the background" message somewhat false.
When this happens though, there must have been a system error for days that we were not able to fix.

GitOrigin-RevId: 33b2661ddbd36379b522fa72c737e48a22df5de4