remove limit on total user indexes (#25065)
there are limits on total number of tables, and limits on indexes per table. this implies a limit on total indexes. this effectively lifts the limit from 256 to 32 * 10000 = 32k

while we're here, let's make the limit on indexes per table enforced in the IndexModel instead of just in the parsing of `TableDefinition` json.

GitOrigin-RevId: 930c06fa93af5925e497154a4bc763e6c2b9624f