pass modules_storage to module source fetch (#24963)
this PR does the plumbing to make the `modules_storage: Arc<dyn Storage>` available in the functions that fetch module source code. Module source code is (still) fetched only from `_module_versions`, but we eventually want it to be fetched from S3 instead.

this PR is piping the storage through the funrun and application structs, and also adding some Debug impls which i found useful while figuring this out.

in particular is the `StorageForInstanceImpl` naming and pattern okay?

GitOrigin-RevId: 3aaeab09e43c64ee76c3ea22bc2197a610363a93