move DeletedBitset to common (#25103)
I plan to reuse `DeletedBitset` for the search deletion tracker because the tantivy is also 4 bytes (same as the point offset). In the next PR, the incremental indexing code is in a `text_search` crate that `search` depends on, and `search` also depends on `vector`, so common seemed like the best thing for both `vector` and `text_search` to depend on.

GitOrigin-RevId: 030be3822c6a08744846bafa4dc54db18414068c