[modules] populate cache with multiple values (#25152)
when we start storing modules in S3, we will want to populate the entire module cache at once by reading the S3 zip file once. To do this, we should add a method to AsyncLru `get_and_prepopulate` that has the same functionality as `get` but the ValueGenerator can return values for additional keys, prepopulating them for future gets.

Started using this method in `ModuleCacheWorker` with no benefits over the existing API -- it's just exercising the codepath in preparation for the S3 use-case.
Also add a unit test.

GitOrigin-RevId: 1191f1558778384cee7690e3f4fb9fcc17ffb07a