allow Date.now in schema evaluation (#25223)
set Date.now to a fixed timestamp determined at the beginning of schema evaluation. We do not store this timestamp anywhere; in particular it's not the same timestamp as used for Analyze. But I think that's fine because evaluating the schema really shouldn't be making decisions that change the schema with `Date.now` calls. Basically it's the same as how we treat Math.random in schema evaluation. Note future plan is described in #18861 .

GitOrigin-RevId: 1e8f02217b0383e329f097d036c5e7afa234288c