Search deleted terms tracker (#25254)
This PR adds the logic for the search deleted terms tracker. We'll write three tracker files - id tracker, deleted bitset, and deleted terms since not all of them are needed everywhere. The deleted terms file format is pretty similar to id tracker but instead of using perfect hashing, we store the term ordinals and number of deleted documents corresponding to that term in separate structures: `EliasFano` for the `TermOrdinal`s because they are monotonic increasing integers, and `DacsOpt` for the counts of deleted documents because they will be small positive integers.

GitOrigin-RevId: 8e42c994af566b27d0a16219898e36459ec19af1