Build tantivy index segment from revision stream (#25383)
This PR adds a function to create a tantivy segment from a stream of document revisions in descending timestamp order. We keep the invariant that convex ids are unique in tantivy segments, so we don't process deletes from the revision stream, we just skip writing them.

I also added tests for delete and replace and fixed a small bug where deleted documents weren't used if you only had OR terms in the search query.

GitOrigin-RevId: 47a940c5ec09454c1c321851a183ec2f503a073b