Add support for HTTP action response streaming (#24763)
Without this change, HTTP actions support streaming APIs (e.g. I can construct a `Response` where the body is a stream), but we'd buffer the entire response body before sending any part of the response.

With this change, we're doing actual streaming -- the isolate sends `HttpResponsePart`s, where the first is the "head" with the status code + headers, and every remaining part is a chunk of the body as bytes.

One of the more complicated parts is how to handle errors -- once we've started streaming, if we hit an error, we can't report that error in the response, but we still want to surface it somehow. In these cases, I'm adding the error as a log line (so it'll appear in the dashboard / log streams / npx convex logs).

But it might also be nice if hitting an error before we've started streaming results in a 500 with the (potentially redacted) error message (our current behavior).

GitOrigin-RevId: 154af575a400c99809d1f2b5ce3f091128ba82ed