nest DeveloperDocument inside ResolvedDocument (#25379)
DeveloperDocument is basically a Value with two fields denormalized out: `_id` and `_creationTime`. Meanwhile ResolvedDocument is a DeveloperDocument that also has a tablet id.
let's structure the types like that.
keep the interface almost the same; we can change it later. this removes the generic and many usages of TableIdAndNumber.

GitOrigin-RevId: b214895e1ec0c725b6da2bd58a4f4e859095d789