Rename Identity proto to UncheckedIdentity (#25495)
This is a no-op refactor that renames Identity to avoid confusion. I am going to introduce IdentityToken in a future PR and would like to decouple the two.

We will likely eventually deprecate this struct in favor of validating the token at the begin timestamp similar to how Node actions.

GitOrigin-RevId: 9c03c5c9da6f460cc07a41c016ef292f6e64ddc3