Get allowed visibility based on the caller instead of passing it seperatly. (#25521)
Allowed visibility is completely derivable by the function caller. There are many places we pass the two alongside each other but this is error prone. The redundant code is especially problematic as we introduce more layers and services.

Instead of this, add a method to functionCaller that gives is allowed visibility, and deleted all redundant code.

GitOrigin-RevId: f8167938fb011f15fab2b175a5e6c6c66e9df2e7