Remove block_logging from ActionCallbacks trait. (#25534)
block_logging is very confusing. It is part of ActionCallbacks trait, which means it is serialized and deserialized from protos.  However, all that obfuscates the fact that block_logging is always false for action callbacks. Thus the whole thing is redundant.

To help clean this up, remove block_logging from the trait and set it to false. We could pass log_visibility into ApplicationFunctionRunner, so we can check if we should block visibility, but this will always return false since actions have AllowedVisibility::All anyway, so passing this and creating transactions in order to do the check seem like an overkill.

GitOrigin-RevId: 3be542a7f2860e31b36a7506be1fa2054bb1b28a