detect race with module version changing (#25556)
previously we were doing an indexed lookup on `_module_versions` which restricted the version. then we stopped doing this. this exposes a rare race condition where the wrong module version can be populated with a cache lookup, if the module changes between the function starting and loading the source from cache.

to fix, we should transactionally look up the version in `_modules` and compare it to the expected version.

GitOrigin-RevId: 27afb616ce056438fcb2b65ec1208d12386cb0d1