Move the alive bitset into the intersection query (#25554)
This change simplifies our tantivy query: Instead of computing the intersection of the filter terms with the union of the query terms and then filtering out all deleted matches, we can just compute the intersection of the filter terms with the alive document IDs and the union of the query terms.

```scheme
(exclude deleted (and (or *query_terms) *filter_terms))
=> (and alive (or *query_terms) *filter_terms)
```

I made some changes to our fork of Tantivy to make their `BitSetDocSet` work with the `ReadOnlyBitSet` that I'll tweak a bit if you like this approach before landing.

GitOrigin-RevId: 1bc8a84076e4bad7dab7549859fbe09e40c352b3