Add ValidatedPathAndArgs proto. (#25743)
Currently, we use common.PathAndArgs proto for two different use cases some are validated and some are not. This is error prone:
- We can create ValidatedPathAndArgs from PathAndArgs proto. This makes it easy to create a bug where we miss validation.
- We construct PathAndArgs with None npm_version in a bunch of places. It is unclear if this is correct.

To solve this, rename PathAndArgs to ValidatedPathAndArgs and add another one without npm_version and use them in the appropriate places. This is a no-op refactor as all fields should be compatible.

GitOrigin-RevId: df84db8074e045332172d28e40c2d7919f42ac43