Redact errors in backend (#25763)
Make execute_query2 use separate protos that make clear the error and log lines are potentially redacted. I debated whether to return redacted protos, or unredacted ones and redact in the Usher. The reason I decided to return redacted is:
1. Due to the specific of the redaction rules, the backend currently is the only one that knows whether it should redact or not. In order for Usher to know, backend needs to tell it, and it makes it simpler to just redact.
2. We redact log for production instances. In that case, it is better for backend to not send logs, instead of backend to send logs only for Usher to throw them away. This optimizes for high load in production.
3. This is already how the code is structured and returning unredacted logs requires big relaying. It is better to not be forced to do that now.

I have kept RedactedJsError map one to one to the current implementation. We can change this separately if we decide to.

GitOrigin-RevId: 2ae8c697921063aa381ff0fbc3acb7bc23315a91