Wire up `ctx.run*` async syscalls to be component aware (#25755)
Wire up `ctx.run*` async syscalls to be component-aware!

There's still some hax that's gated on an environment variable that we'll remove when we (1) namespace modules and (2) add the reference resolution algorithm.

But, a basic v8 action is able to call into another "component" with `Convex.asyncSyscall`!
```ts
export const demo = action({
  handler: async (ctx, args) => {
    if (args.stop) {
      console.log("done!");
      return;
    }
    const argsJSON = {
      reference: "_reference/childComponent/waitlist/messages/demo",
      args: convexToJson({ stop: true }),
      version: "1.11.3",
      requestId: "",
    };
    const resultStr = await Convex.asyncSyscall(
      "1.0/actions/action",
      JSON.stringify(argsJSON),
    );
    const result = JSON.parse(resultStr);
    console.log("hi", result);
  },
});
```

GitOrigin-RevId: 640dcd101ce0794269262becb8a41b792395602d