modules are scoped by ComponentDefinition, not Component (#25796)
modules are defined inside a component, so when we look up a module or refer to it by path, we should do this within the context of a `ComponentDefinition`, not a `Component`.

note that `CanonicalizedComponentFunctionPath` remains scoped by `Component`, because it is used primarily when there is a function to execute. Therefore when converting from `CanonicalizedComponentFunctionPath` to `CanonicalizedComponentModulePath`, e.g. to look up a function for execution, we need to use `ComponentsModel` to do the conversion.

GitOrigin-RevId: 3f2237c23e91afebc8f0b7fb415c017c880330f3