move `index_range` out of UserFacingModel (#25812)
UserFacingModel is for user tables, so it should be scoped to a component.
`index_range_batch` started off as being only for user tables, but recently we started using it for system tables as well, which might not be scoped to a component. let's pull it out since it's common code, not necessarily UserFacing anymore.

GitOrigin-RevId: b4f3342c2e02187c75ec85159ab7b22cccf15aaf