thread ComponentId through UserFacingModel (#25818)
UserFacingModel always deals with user tables, so it's scoped to a component.

the main benefit of this is in tests we can use `new_root_for_tests`, and this will be passed down to the transaction methods underneath, when converting table name and number to tablet. this can also fetch the component ID when we remove the `Root` shortcut.

GitOrigin-RevId: 97222d63b48dea7511c5ef6447a88ec26f693826