Move some funruns protos to common (#25837)
The respective protos are used from open source creates and we want to make funrun.proto and backend.proto to private.

One confusing thing is that we have QueryJournal and Cursor in both common and funrun but they are slightly different and incompatible. I have moved the ones from funrun to common and prefixed the ones already there with Instance and added a common of how they are used.

GitOrigin-RevId: 738b49e0597135be604c6699e596ec082c6bc296