Switch to message instead of messages for console log stream events (#25843)
This makes the easy case of grepping for a certain string easier, and makes the more advanced case of wanting structured log lines slightly harder (probably hooking up a JSON parser in Axiom / Datadog).

I also added a could other fields that seemed useful.

Note that this is in the yet to be launched format, so this isn't a breaking change.

GitOrigin-RevId: 7e82459847e7985e873767798e75295d1f85a68e