Add a v2 of the text index flusher guarded by a knob. (#25861)
I haven't added tests because the interfaces are basically
non-functional. But this at leaves gives us the scaffolding to start
filling things out.

The write metadata part here is a bit throwaway, that'll end up being
handled by the Writer. I just haven't generalized it yet. For now this
is ok so long as we either don't support compaction, or we make
compaction explicitly not write its updates if a flush happened during
compaction.

I'll rename some more search things in a followup

GitOrigin-RevId: 8ad48d73ad637b05497f44848d4ea07a78be76d3